test: fix useIframeBehaviour hook tests
